Extract Flow service and handler interfaces, refactor route handling, and add unit tests for flow route registration.
